refactor : TokenInfo 클래스 -> TokenRequest, TokenResponse로 분리
