+++ src/main/java/com/WearWeather/wear/domain/auth/dto/request/RefresehTokenRequest.java
@@ -1,4 +1,4 @@
-package com.WearWeather.wear.domain.auth.entity;
+package com.WearWeather.wear.domain.auth.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,8 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class TokenInfo {
+public class RefresehTokenRequest {
 
-    private String accessToken;
     private String refreshToken;
 }
