+++ src/main/java/com/WearWeather/wear/domain/auth/dto/response/TokenResponse.java
@@ -1,0 +1,18 @@
+package com.WearWeather.wear.domain.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class TokenResponse {
+
+    private String accessToken;
+    private String refreshToken;
+}
